+++ Tests/ImageFetcherTests/Helpers/Extensions/Networking+Mock.swift
@@ -1,0 +1,45 @@
+//
+//  Networking+Mock.swift
+//  Mobelux
+//
+//  MIT License
+//
+//  Copyright (c) 2023 Mobelux LLC
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import Foundation
+import ImageFetcher
+
+extension Networking {
+    static func mock(
+        responseDelay: TimeInterval? = nil,
+        responseProvider: @escaping (URL) throws -> Data = { _ in Data() }
+    ) -> Self {
+        .init(
+            load: { request in
+                if let responseDelay {
+                    try await Task.sleep(nanoseconds: UInt64(responseDelay * 1_000_000_000))
+                    try Task.checkCancellation()
+                }
+                return try responseProvider(request.url!)
+            })
+    }
+}
